change: suppress warning log message when trying to remove a non-existing marker file
SVN: 7012

+++ datamover/source/java/ch/systemsx/cisd/datamover/DataMover.java
@@ -170,6 +170,10 @@ public final class DataMover
         for (String fileName : PROCESS_MARKER_FILENAMES)
         {
             File markerFile = new File(fileName);
+            if (markerFile.exists() == false)
+            {
+                continue;
+            }
             if (markerFile.delete() == false)
             {
                 operationLog.warn("Couldn't delete process marker file "
